Add Patch&Delete Permission, breaks Add Permission
Spent all day trying to make this this work. Was able to add the ability
to Change Permissions for a given Box (change read/write status), and
delete Permission. In order to do this, I had to establish the
Users<>UsersBoxes<>Boxes relationships, but something about having that
relationship in the models is breaking the ability to create a new
UsersBox. Issue #2